requirements.txt
Pyrogram==2.0.106
python-dotenv==1.0.0
TgCrypto==1.2.5